Pull file loading and circle generation out of Epicycles
