+++ shapes/H.txt
@@ -1,4 +1,3 @@
-shape
 -1.5 -1.5
 -1.45238095238095 -1.5
 -1.4047619047619 -1.5
