+++ shapes/closed_hilbert.txt
@@ -1,4 +1,3 @@
-shape
 -0.5 -7.5
 -0.547619047619048 -7.5
 -0.595238095238095 -7.5
