--- shapes/diamond.txt
@@ -1,7 +1,0 @@
-harmonics
-250 1j
-27.8 -3j
-10 5j
-5.1 -7j
-3.1 9j
-2.1 -11
