+++ shapes/heart.txt
@@ -1,4 +1,3 @@
-shape
 0 7.5370664958657
 0.000503365998346545 7.55748296405302
 0.00402090979321222 7.61849185098036
