--- shapes/square_wave.txt
@@ -1,6 +1,0 @@
-harmonics
-180 1j
-60 3j
-36 5j
-25.7 7j
-20 9j
