+++ shapes/star.txt
@@ -1,4 +1,3 @@
-shape
 0.951056516295154 0.213525491562421
 0.93222371399228 0.213525491562421
 0.913390911689405 0.213525491562421
