+++ shapes/triangle.txt
@@ -1,4 +1,3 @@
-shape
 -0.5 -0.433
 -0.480392156862745 -0.433
 -0.46078431372549 -0.433
