+++ src/constants.py
@@ -3,7 +3,7 @@ FPS = 60
 DT_LIMIT = 2 / FPS  # max dt is twice the normal dt
 DEFAULT_SCALE_FACTOR = 0.8
 PATH_COLOR = (255, 0, 0)
-BACKGROUND_COLOR = (255, 255, 255)
+BACKGROUND_COLOR = (0, 255, 255)
 CIRCLE_COLOR = (170, 170, 170)
 DEFAULT_ANGULAR_VELOCITY = 1  # in radians/second
 MIN_ANGULAR_VELOCITY = 1 / 32
