+++ src/image_loader.py
@@ -1,0 +1,5 @@
+# TODO: Load an image file. Ideally the path should be a single pixel wide
+#  and must be of a single color. Construct the point list either by using
+#  nearest neighbor search or try finding the shortest path using
+#  a travelling salesman approach.
+# Return the sorted list of coordinates.
